requirements.txt
gradio==4.44.1
torch
torchvision
requests
Pillow
numpy